requirements.txt
APScheduler==3.6.3
atomicwrites==1.4.0
attrs==21.4.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
idna==3.3
iniconfig==1.1.1
isort==5.10.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
python-dotenv==0.19.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
tomli==2.0.1
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
flake8==3.9.2